Moving handlers logic to handlers class
